+++ inegraphics/src/main/java/com/inepex/inegraphics/shared/gobjects/MoveTo.java
@@ -3,7 +3,7 @@ package com.inepex.inegraphics.shared.gobjects;
 public class MoveTo extends PathElement {
 
 
-	public MoveTo(int endPointX, int endPointY) {
+	public MoveTo(double endPointX, double endPointY) {
 		super(endPointX, endPointY);
 	}
